+++ src/imephu/salt/annotation/salticam.py
@@ -1,0 +1,37 @@
+from astropy import units as u
+from astropy.coordinates import SkyCoord
+from astropy.wcs import WCS
+
+from imephu.annotation.general import CircleAnnotation, GroupAnnotation, TextAnnotation
+
+
+def field_of_view_annotation(fits_center: SkyCoord, wcs: WCS) -> GroupAnnotation:
+    """
+    Return an annotation with the Salticam field of view.
+
+    Parameters
+    ----------
+    fits_center: `~astropy.coordinates.SkyCoord`
+        The central position of the finder chart, in right ascension and declination.
+    wcs: `~astropy.wcs.WCS`
+        WCS object.
+
+    Returns
+    -------
+    `~imephu.annotation.general.GroupAnnotation`
+        An annotation with the Salticam field of view.
+    """
+    fov_annotation = CircleAnnotation(
+        fits_center, 5 * u.arcmin, wcs=wcs, edgecolor="green"
+    )
+    name_annotation = TextAnnotation(
+        (0.86, 0.86),
+        "SCAM",
+        wcs=wcs,
+        style="italic",
+        weight="bold",
+        size="large",
+        horizontalalignment="left",
+        color=(0, 0, 1),
+    )
+    return GroupAnnotation([fov_annotation, name_annotation])
