+++ tests/imephu/salt/annotation/test_rss/test_rss_field_of_view_annotation.png
