+++ tests/imephu/salt/annotation/test_salticam/test_salticam_field_of_view_annotation.png
